+++ civis/polling.py
@@ -27,6 +27,10 @@ class PollableResult(CivisAsyncResultBase):
     api_key : str, optional
         This is not used by PollableResult, but is required to match the
         interface from CivisAsyncResultBase.
+    poll_on_creation : bool, optional
+        If ``True`` (the default), it will poll upon calling ``result()`` the
+        first time. If ``False``, it will wait the number of seconds specified
+        in `polling_interval` from object creation before polling.
 
     Examples
     --------
@@ -59,11 +63,21 @@ class PollableResult(CivisAsyncResultBase):
     # - We use the `Future` thread lock called `_condition`
     # - We assume that results of the Future are stored in `_result`.
     def __init__(self, poller, poller_args,
-                 polling_interval=_DEFAULT_POLLING_INTERVAL, api_key=None):
-        super().__init__(poller, poller_args, polling_interval, api_key)
+                 polling_interval=None, api_key=None,
+                 poll_on_creation=True):
+        if polling_interval is None:
+            polling_interval = _DEFAULT_POLLING_INTERVAL
+        super().__init__(poller,
+                         poller_args,
+                         polling_interval,
+                         api_key,
+                         poll_on_creation)
 
         # Polling arguments. Never poll more often than the requested interval.
-        self._last_polled = None
+        if poll_on_creation:
+            self._last_polled = None
+        else:
+            self._last_polled = time.time()
         self._last_result = None
 
         self._self_polling_executor = None
@@ -123,16 +137,24 @@ class PollableResult(CivisAsyncResultBase):
                     # If the job has finished, then register completion and
                     # store the results. Because of the `if self._result` check
                     # up top, we will never get here twice.
-                    if self._last_result.state in FAILED:
-                        try:
-                            err_msg = str(self._last_result['error'])
-                        except:
-                            err_msg = str(self._last_result)
-                        self.set_exception(CivisJobFailure(err_msg,
-                                           self._last_result))
-
-                        self._result = self._last_result
-                    elif self._last_result.state in DONE:
-                        self.set_result(self._last_result)
+                    self._set_api_result(self._last_result)
 
             return self._last_result
+
+    def _set_api_result(self, result):
+        with self._condition:
+            if result.state in FAILED:
+                try:
+                    err_msg = str(result['error'])
+                except:
+                    err_msg = str(result)
+                self.set_exception(CivisJobFailure(err_msg, result))
+                self._result = result
+                self.cleanup()
+            elif result.state in DONE:
+                self.set_result(result)
+                self.cleanup()
+
+    def cleanup(self):
+        # This gets called after the result is set
+        pass